# docs/swagger.yaml
basePath: /api/v1
definitions:
  Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message: {}
    type: object
  auth.CreateUserRequest:
    properties:
      degree:
        type: string
      email:
        type: string
      gender:
        type: string
      nickname:
        type: string
      password:
        type: string
      password_again:
        type: string
      phone_num:
        type: string
    required:
    - degree
    - email
    - gender
    - nickname
    - password
    - password_again
    - phone_num
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  errno.Errno:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  forum.CommentRequest:
    properties:
      content:
        type: string
      postID:
        type: integer
    type: object
  forum.ForumRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  model.Confirm:
    properties:
      accepter:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      publisher:
        type: string
      taskID:
        type: integer
      updatedAt:
        type: string
    type: object
  model.Task:
    properties:
      accepter:
        type: string
      award:
        type: integer
      confirm:
        $ref: '#/definitions/model.Confirm'
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      image:
        items:
          $ref: '#/definitions/model.TaskImage'
        type: array
      prefecture_id:
        type: string
      publisher:
        type: string
      status:
        type: string
      tag:
        type: string
      updatedAt:
        type: string
      willing:
        items:
          $ref: '#/definitions/model.Willing'
        type: array
      willing_num:
        type: integer
    type: object
  model.TaskImage:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      path:
        type: string
      sha:
        type: string
      taskID:
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
  model.Willing:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      taskID:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: 119.3.2.168:2333
info:
  contact:
    email: 1903180340@qq.com
    name: Wishiforpeace
  description: 纸贵API
  termsOfService: http://swagger.io/terrms/
  title: Zhigui API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 邮箱登录
      parameters:
      - description: 注册用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"将student_id作为token保留"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"error_code":"20001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/errno.Errno'
        "401":
          description: '{"error_code":"10001", "message":"The email address has been
            registered"} '
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 邮箱注册登录
      parameters:
      - description: 注册用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/auth.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"将student_id作为token保留"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"error_code":"20001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/errno.Errno'
        "401":
          description: '{"error_code":"10001", "message":"The email address has been
            registered"} '
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: Register
      tags:
      - auth
  /forum:
    get:
      consumes:
      - json/application
      description: 分页查看全部帖子
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'limit--偏移量指定开始返回记录之前要跳过的记录数 '
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 查看帖子
      tags:
      - forum
  /forum/comments:
    get:
      consumes:
      - json/application
      description: 查看已发布帖子的评论内容
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: id--帖子的id
        in: query
        name: id
        required: true
        type: integer
      - description: 'limit--偏移量指定开始返回记录之前要跳过的记录数 '
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取某帖子评论
      tags:
      - forum
    post:
      consumes:
      - json/application
      description: 用户根据需要发布帖子
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: id--任务的id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forum.CommentRequest'
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 发布帖子
      tags:
      - forum
  /forum/like:
    post:
      consumes:
      - json/application
      description: 查看已发布帖子的评论内容
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子的id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取某帖子评论
      tags:
      - forum
...
  /forum/publish:
    post:
      consumes:
      - json/application
      description: 用户根据需要发布帖子
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: id--任务的id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forum.ForumRequest'
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 发布帖子
      tags:
      - forum
  /forum/publish/pictures:
    post:
      consumes:
      - json/application
      description: 上传新的图片
      parameters:
      - description: 获取email
        in: header
        name: Authorization
        required: true
        type: string
      - description: id--帖子的id
        in: formData
        name: id
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - json/application
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 上传帖子图片
      tags:
      - forum
...
schemes:
- http
swagger: "2.0"